Homework PUT method added(not working)

+++ src/main/java/ie/atu/week2/week2_CICD2/Product.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 public class Product {
     public long id;
     public String name;
-    public long price;
+    public double price;
 }
